+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ## Added
 - registry: Experimental PostgreSQL metadata database (incomplete, in progress)
 - registry: Use GitLab LabKit for HTTP metrics collection
+- registry: Add new configuration for changing the output for logs and the access logs format
+- registry: Use GitLab LabKit for correlation and logging
+- registry: Normalize log messages
 
 ## [v2.9.1-gitlab] - 2020-05-05
 ## Added